+++ CHANGELOG.md
@@ -5,6 +5,13 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 -----
 
+## [1.2.0] - 2017-03-19
+
+**Changed:**
+
+- Add `Response.Status()`, `Response.Type()`.
+- `Context.Status` and `Context.Type` will not return value now.
+
 ## [1.1.3] - 2017-03-15
 
 **Changed:**